[Feat] 몬스터 체력바 구현

+++ Assets/Scripts/UI/CharacterView/UI_HealthBar.cs
@@ -5,7 +5,7 @@ public class UI_HealthBar : MonoBehaviour
 {
     [SerializeField] private Slider slider;
 
-    public void UpdateHealthBar(int curHP, int maxHP)
+    public void UpdateHealthBar(float curHP, float maxHP)
     {
         slider.maxValue = maxHP;
         slider.value = curHP;
